[cross-ctrl-vldn] Create custom validators for the new pwd
The validator fns are passed the control & return either
null (if there are no errors) or an error object.
The key-value pairs in this error object are  merged into
the errors object on the control. Therefore, the hasError method
can be used with the key set here to check if the validation failed.