[cross-ctrl-vldn] Check that new pwd doesn't match old pwd
We attach a validator on the FormGroup, so we can get access to the
child controls. In the function, we compare the value in one control
with that in another.

Instead of returning an error object for the FormGroup, we manually
toggle the error on the new pwd control. Since we manage the error
manually, we must be careful to unset it when the validation passes.